Fix async behavior that broke SDK tests
When async task throws exception that exception is always wrapped in AggredatedException. The original sync code didn't throw that exception type and thar broke the interface of exceptions the use is catching. To fix that all sync methods that invoke async methods internally not call a method that either return the result or re-throw the actual exception instead of the aggregated exception.

Replace usage of RunSynchronously with Wait RunSynchronously is not allowed for the task type we are using.

Fix GetWithGroupByAsync because Octane server was changed to return logical names by default but the SDK is still expecting numeric id.