Fixed #4 - Switch BaseEntity ID to be string while still supporting int for now.
Octane switch to using string as EntityId but since this SDK use long we are using a new type call EntityId that support implicit convertion to and from long and string.
There is Obsolete notice when using the long conversions which hopefully cause the SDK developers to switch to string.